refacto: Move _add_labels to asbtract concept resolver
