+++ app/services/concepts/idref_concept_solver.py
@@ -6,7 +6,6 @@ from aiohttp import ClientTimeout
 from rdflib import Graph, SKOS
 
 from app.db.models.concept import Concept as DbConcept
-from app.db.models.label import Label as DbLabel
 from app.services.concepts.concept_solver import ConceptSolver
 from app.services.concepts.dereferencing_error import DereferencingError
 
@@ -62,48 +61,6 @@ class IdRefConceptSolver(ConceptSolver):
             raise DereferencingError(
                 f"Unknown error while dereferencing {idref_url} with message {error}"
             ) from error
-
-    def _add_labels(
-        self,
-        concept: DbConcept,
-        labels: list[rdflib.term.Literal],
-        preferred: bool = True,
-    ):
-        def interesting_labels(label):
-            return (
-                label.language in self.settings.concept_languages
-                or label.language is None
-            )
-
-        if len(labels) == 0:
-            return
-        # If there is only one preflabel, add it
-        # If there are several preflabels, and one one them
-        # is in self.settings.concept_languages or has no language,
-        # add only preflabels whose language is in self.settings.concept_languages
-        # and preflabels whithout language
-        # If there are several preflabels, and none of them
-        # is in self.settings.concept_languages or has no language,
-        # add the first one in any language
-        if preferred:
-            preferred_labels = [label for label in labels if interesting_labels(label)]
-            for label in preferred_labels or [labels[0]]:
-                self._add_label(concept, label, preferred)
-        # Add all altlabels whose language is in self.settings.concept_languages
-        # and altlabels whithout language
-        else:
-            alt_labels = [label for label in labels if interesting_labels(label)]
-            for label in alt_labels:
-                self._add_label(concept, label, preferred)
-
-    def _add_label(self, concept, label, preferred):
-        concept.labels.append(
-            DbLabel(
-                value=str(label),
-                language=label.language,
-                preferred=preferred,
-            )
-        )
 
     async def _build_url_from_concept_id_or_uri(self, concept_id):
         original_concept_id = concept_id
